Making syncSupervisors its own thread, with callers notifying it to do work through a TVar.
This alleviates a race condition where it was possible to launch a given process two times,
once through a call to syncSupervisors from monitorConfig and once from a call to
syncSupervisors from pollStale.  The running field of the sharedGroupConfig would only end
up knowing about one of the running processes, but there would be two of them with two
seperate supervise threads watching them.