Adding minRestartDelay configuration.
If minRestartDelay > delay, then it has the capacity to increase the amount of
time spent waiting before restarting a monitored process.  The supervise thread
monitoring the process will delay at least `delay` seconds between the
termination of the process and restarting.  It will delay at least
`minRestartDelay` seconds between the _starting_ of the process and the
next restarting of the process after its termination.

This allows supervised processes to set a delay of `0`, but avoid continuously
restarting the process as fast as possible if some configuration error causes
the process to not start successfully.